Reorganize components files into its own folder
