+++ src/components/composed/Cover/Cover.js
@@ -1,0 +1,11 @@
+
+import {CoverImg} from '../../../assets/images'
+
+
+const Cover = () => {
+    return(
+        // eslint-disable-next-line jsx-a11y/alt-text
+        <div><img width='500px'src={CoverImg} a1t='Logo and slogan' /> </div> 
+    )}
+
+export default Cover;
