+++ src/components/composed/Footer/Footer.js
@@ -1,0 +1,12 @@
+
+import SocialMedia from '../SocialMedia/SocialMedia'
+import {FooterWrapper} from './Footer.style'
+
+const FooterC = () => {
+    return(
+        <FooterWrapper>      
+            <SocialMedia />
+        </FooterWrapper>         
+    )
+}
+export default FooterC;
