+++ src/components/composed/SocialMedia/SocialMedia.js
@@ -1,0 +1,14 @@
+
+
+const SocialMedia = () => {
+    return(
+        <div direction='row' justify='evenly' gap='medium' >
+            <a target='_blank' rel='noreferrer' href='https://twitter.com' >Twitter</a>
+            <a target='_blank' rel='noreferrer' href='https://www.instagram.com'>Instagram</a>
+            <a target='_blank' rel='noreferrer' href='https://www.facebook.com/'>Facebook </a>
+        </div>
+        
+    )
+}
+
+export default SocialMedia;
